fix(e2e): guard on mapped sources, move the inline flag to the build call
Round two found the empty-report guard blind to the case its own message
described: a bundle whose sourcemap can't be found survives as a report file
under its own dist path, so the length check passed and Codecov would have
been handed coverage for a file that doesn't exist in the repo. The guard now
demands every file resolve under src/ludamus/client/src.

The inline-sourcemap decision moves out of vite.config and onto the one call
that wants it — `test:e2e:prep` passes `--sourcemap inline`. The config goes
back to declaring `sourcemap: true` with no environment sniffing, so shipped
builds emit their own maps and nothing about a build depends on a variable
named for coverage data.

The config also defaults COVERAGE_FILE before building the server env, so a
runner started outside mise agrees with the server it spawns instead of
silently collecting nothing. helpers/test.ts becomes helpers/fixtures.ts: a
file called test.ts inside testDir invites someone to "fix" it later.
