+++ src/ludamus/client/vite.config.ts
@@ -66,9 +66,7 @@ export default defineConfig({
         timetable: resolve(rootDir, "src/timetable.ts"),
       },
     },
-    // Coverage runs need the map inside the bundle: V8 hands back the source
-    // text, and that is what carries the coverage home to the .ts files.
-    sourcemap: process.env.COVERAGE_FILE?.endsWith(".coverage.e2e") ? "inline" : true,
+    sourcemap: true,
   },
   plugins: [djangoTemplateReload(), tailwindcss()],
   server: {
