+++ tests/e2e/playwright.config.ts
@@ -28,12 +28,15 @@ const WEB_COMMAND =
 const isCI = !!process.env.CI;
 const skipIos = !!process.env.E2E_SKIP_IOS;
 
+// Set before webServerEnv is built, so the runner and the server it starts
+// agree on whether this run measures coverage.
+process.env.COVERAGE_FILE ??= path.join(repoRoot, ".coverage.e2e");
+
 const webServerEnv: Record<string, string> = Object.fromEntries(
   Object.entries(process.env).filter(
     (entry): entry is [string, string] => typeof entry[1] === "string",
   ),
 );
-webServerEnv.COVERAGE_FILE ??= path.join(repoRoot, ".coverage.e2e");
 
 export default defineConfig({
   testDir: "./tests",
