+++ tests/e2e/tests/anonymous-proposal.spec.ts
@@ -1,4 +1,4 @@
-import { expect, test } from "./helpers/test";
+import { expect, test } from "./helpers/fixtures";
 
 test.describe("Anonymous proposals", () => {
   test("redirects to login when anonymous proposals are disabled", async ({ page }) => {
