+++ tests/e2e/tests/event-schedule-rail.spec.ts
@@ -1,4 +1,4 @@
-import { expect, test } from "./helpers/test";
+import { expect, test } from "./helpers/fixtures";
 
 const denseEventUrl = "/chronology/event/kapitularz-2025-anonymized/";
