+++ tests/e2e/tests/flatpages.spec.ts
@@ -1,4 +1,4 @@
-import { expect, test } from "./helpers/test";
+import { expect, test } from "./helpers/fixtures";
 
 test.describe("Flatpages", () => {
   test("About page displays seeded content", async ({ page }) => {
