+++ tests/e2e/tests/panel.spec.ts
@@ -2,7 +2,7 @@ import { type Locator, type Page } from "@playwright/test";
 import fs from "node:fs";
 import path from "node:path";
 
-import { expect, test } from "./helpers/test";
+import { expect, test } from "./helpers/fixtures";
 
 /** Accept the in-page confirm modal that guards destructive forms. */
 const acceptConfirmModal = (page: Page) =>
