+++ tests/e2e/tests/profile.auth.spec.ts
@@ -1,4 +1,4 @@
-import { expect, test } from "./helpers/test";
+import { expect, test } from "./helpers/fixtures";
 
 test.describe("Profile — Personal Information (edit.html)", () => {
   test("shows profile form with tab navigation", async ({ page }) => {
