+++ tests/e2e/tests/promotion.auth.spec.ts
@@ -1,7 +1,7 @@
 import fs from "node:fs";
 import path from "node:path";
 
-import { expect, test } from "./helpers/test";
+import { expect, test } from "./helpers/fixtures";
 
 // Multi-user waiting-list promotion (issue scenario 1): the superuser (A) holds
 // the only confirmed seat and a dedicated waiter (B) is waitlisted. When A
