+++ tests/e2e/tests/timetable.spec.ts
@@ -1,6 +1,6 @@
 import { type Page } from "@playwright/test";
 
-import { expect, test } from "./helpers/test";
+import { expect, test } from "./helpers/fixtures";
 
 test.describe.configure({ mode: "serial" });
